// src/wardrobe/garment-category.enum.ts
export enum GarmentCategory {
  TOPS = 'tops',
  BOTTOMS = 'bottoms',
  DRESSES = 'dresses',
  OUTERWEAR = 'outerwear',
  FOOTWEAR = 'footwear',
  ACCESSORIES = 'accessories',
  BAGS = 'bags',
  ACTIVEWEAR = 'activewear',
  SWIMWEAR = 'swimwear',
  UNDERWEAR = 'underwear',
  LINGERIE = 'lingerie',
  OTHER = 'other',
}